Refactor: #12 viewmodel, model 파일 생성해서 view와 분리, 리팩토링
