+++ Music-Player/Music-Player/Model/NowPlayingSong.swift
@@ -1,0 +1,16 @@
+//
+//  NowPlayingSong.swift
+//  Music-Player
+//
+//  Created by Junhong Park on 2022/01/16.
+//
+
+import SwiftUI
+
+struct NowPlayingSong {
+    var title: String
+    var albumTitle: String
+    var artist: String
+    var artWork: UIImage
+    var totalRate: Double
+}
